refactoring folder structure
