+++ src/app/app-routing.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { ShoppingListComponent } from './shopping-list/shopping-list.component';
-import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
-import { RecipeComponent } from './recipe/recipe.component';
-import { HomeComponent } from './home/home.component';
+import { ShoppingListComponent } from './components/shopping-list/shopping-list.component';
+import { PageNotFoundComponent } from './components/page-not-found/page-not-found.component';
+import { RecipeComponent } from './components/recipe/recipe.component';
+import { HomeComponent } from './components/home/home.component';
 
 
 const appRoute: Routes = [
