+++ src/app/components/recipe/recipe-detail/recipe-detail.component.html
