+++ src/app/components/recipe/recipe-detail/recipe-detail.component.ts
@@ -1,10 +1,11 @@
 import { Component, OnInit, Input, ViewChild, OnDestroy, TemplateRef } from '@angular/core';
 
-import { RecipeModel } from '../recipe.model';
-import { ModalService } from '../modal.service';
+import { RecipeModel } from '../../../shared/models/recipe.model';
 import { BsModalService, BsModalRef } from 'ngx-bootstrap/modal';
-import { RecipeService } from '../recipe.service';
-import { shoppingItemModel } from 'src/app/shopping-list/shopping-item.model';
+
+import { RecipeService } from '../../../services/recipe.service';
+import { ModalService } from '../../../services/modal.service';
+import { shoppingItemModel } from '../../../shared/models/shopping-item.model';
 
 @Component({
   selector: 'app-recipe-detail',
