+++ src/app/components/recipe/recipe-list/recipe-item/recipe-item.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { RecipeModel } from '../../recipe.model';
-import { ModalService } from '../../modal.service';
+import { RecipeModel } from '../../../../shared/models/recipe.model';
+import { ModalService } from '../../../../services/modal.service';
 
 @Component({
   selector: 'app-recipe-item',
