+++ src/app/components/recipe/recipe.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { RecipeModel } from './recipe.model';
 
 @Component({
   selector: 'app-recipe',
