+++ src/app/components/shopping-list/shopping-item/shopping-item.component.css
