+++ src/app/components/shopping-list/shopping-item/shopping-item.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { shoppingItemModel } from '../shopping-item.model';
+import { shoppingItemModel } from '../../../shared/models/shopping-item.model';
 @Component({
   selector: 'app-shopping-item',
   templateUrl: './shopping-item.component.html',
