+++ src/app/shared/models/shopping-item.model.ts
